# api-getaway/api/docs/swagger.yaml
definitions:
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  orders.CreateOrderRequest:
    properties:
      delivery_address:
        type: string
      delivery_time:
        type: string
      items_id:
        type: string
      kitchen_id:
        type: string
      total_amount:
        type: number
      user_id:
        type: string
    type: object
  orders.CreateOrderResponse:
    properties:
      order:
        $ref: '#/definitions/orders.Order'
    type: object
...
  orders.GetKitchenOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/orders.Order'
        type: array
    type: object
...
  orders.GetOrderByIDResponse:
    properties:
      order:
        $ref: '#/definitions/orders.Order'
    type: object
...
  orders.GetUserOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/orders.Order'
        type: array
    type: object
  orders.Order:
    properties:
      created_at:
        type: string
      delivery_address:
        type: string
      delivery_time:
        type: string
      id:
        type: string
      items_id:
        type: string
      kitchen_id:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  orders.Status:
    properties:
      status:
        type: boolean
    type: object
...
  orders.UpdateOrderStatusRequest:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
...
  users.GetProfileResponse:
    properties:
      user:
        $ref: '#/definitions/users.User'
    type: object
...
  users.Status:
    properties:
      Status:
        type: boolean
    type: object
...
  users.UpdateProfileRequest:
    properties:
      UserId:
        type: string
      address:
        type: string
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      is_verified:
        type: boolean
      password:
        type: string
      phone:
        type: string
      specialties:
        type: string
      user_type:
        type: string
      username:
        type: string
      years_of_experience:
        type: integer
    type: object
  users.User:
    properties:
      address:
        type: string
      bio:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_verified:
        type: string
      password:
        type: string
      phone_number:
        type: string
      specialties:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
      username:
        type: string
      years_of_experience:
        type: integer
    type: object
host: localhost:4444
info:
  contact: {}
  description: This is a sample server for Auth Service.
  title: Auth Service API
  version: "1.0"
paths:
...
  /api/v1/orders:
    post:
      consumes:
      - application/json
      description: Create Order
      parameters:
      - description: Create Order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/orders.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.CreateOrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Create Order
      tags:
      - Orders
  /api/v1/orders/{id}:
    get:
      consumes:
      - application/json
      description: Get Order By Id
      parameters:
      - description: Get Order By Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.GetOrderByIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Order By Id
      tags:
      - Orders
  /api/v1/orders/kitchen/{id}:
    get:
      consumes:
      - application/json
      description: List Kitchen Orders
      parameters:
      - description: List Kitchen Orders
        in: path
        name: id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.GetKitchenOrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: List Kitchen Orders
      tags:
      - Orders
  /api/v1/orders/status:
    put:
      consumes:
      - application/json
      description: Update Order Status
      parameters:
      - description: Update Order Status
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/orders.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.Status'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      tags:
      - Orders
  /api/v1/orders/user/{id}:
    get:
      consumes:
      - application/json
      description: Get User Orders
      parameters:
      - description: Get User Order
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.GetUserOrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get User Orders
      tags:
      - Orders
...
  /api/v1/users/profile/{id}:
    delete:
      consumes:
      - application/json
      description: delete profile
      parameters:
      - description: Delete Profile
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.Status'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: delete profile
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get Profile
      parameters:
      - description: Get Profile
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get Profile
      tags:
      - users
  /api/v1/users/profile/update:
    put:
      consumes:
      - application/json
      description: update profile
      parameters:
      - description: Update Profile
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.GetProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: update profile
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"